some fixes. LED flush for debug.

+++ keymap.h
@@ -5,7 +5,7 @@
 #include "usbkeycodes.h"
 
 int get_layer(void);
-uint8_t get_keycode(int layer, uint8_t row, uint8_t col);
+uint8_t get_keycode(int layer, int row, int col);
 
 #define MATRIX_ROWS 9
 #define MATRIX_COLS 8
